Move types to rlang-types.h
